Added new folder structure and feature toggle
